+++ magda-web-client/src/Components/Publisher/PublisherDetails.js
@@ -3,7 +3,6 @@ import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import { Link } from "react-router-dom";
 import ErrorHandler from "../../Components/ErrorHandler";
-import { config } from "../../config";
 import ReactDocumentTitle from "react-document-title";
 import {
     fetchPublisherIfNeeded,
@@ -69,7 +68,9 @@ class PublisherDetails extends Component {
 
         return (
             <ReactDocumentTitle
-                title={`${publisher.name} | Organisations | ${config.appName}`}
+                title={`${publisher.name} | Organisations | ${
+                    this.props.strings.applicationName
+                }`}
             >
                 <div className="publisher-details">
                     <div>
@@ -162,7 +163,11 @@ class PublisherDetails extends Component {
         ) {
             return (
                 <ReactDocumentTitle
-                    title={this.props.publisher.name + " | " + config.appName}
+                    title={
+                        this.props.publisher.name +
+                        " | " +
+                        this.props.strings.applicationName
+                    }
                 >
                     {this.renderContent()}
                 </ReactDocumentTitle>
@@ -204,7 +209,8 @@ function mapStateToProps(state: Object, ownProps: Object) {
         location,
         searchPublisherName,
         hitCount,
-        error
+        error,
+        strings: state.content.strings
     };
 }