+++ CHANGES.md
@@ -3,6 +3,7 @@
 -   Added configurable argument to `magda-web-server` module to accept Google Analytics IDs
 -   Make home page items come from content api
 -   Make header menu items come from content api
+-   Remove all global references to config.appName and allow getting this configuration item from content api
 -   Changed some existing content ids to group things better
 -   Reorganised content api to require less number of requests to get small JSON content items
 -   Email Templates are configurable from content API